requirements.txt
pyyaml==3.11
piexif==1.0.2
